requirements.txt
Django==5.2
djangorestframework==3.16.0
django-filter==25.1

python-dotenv==1.1.0

psycopg2==2.9.10

pillow==11.2.1

phonenumbers==9.0.4
django-phonenumber-field==8.1.0
